Don't execute the rest of batch actions if the user is using Dex and does not have access to anything
